+++ tests/app/idp/idp/urls.py
@@ -21,7 +21,9 @@ from django.views.generic import TemplateView
 
 
 urlpatterns = [
-    path('', TemplateView.as_view(template_name='home/index.html'), name='home'), # Maps the root URL to your home_view
+    path(
+        "", TemplateView.as_view(template_name="home/index.html"), name="home"
+    ),  # Maps the root URL to your home_view
     path("admin/", admin.site.urls),
     path("o/", include("oauth2_provider.urls", namespace="oauth2_provider")),
     path("accounts/", include("django.contrib.auth.urls")),